*Fix: fixing powx issues of r50

+++ inc/gMathFunc.h
@@ -7,7 +7,7 @@ long long toLong(long long x);
 long long toLong(long double x);
 long long sqrtx(long long x);
 long double sqrtx(long double x);
-long double powx(long double x, long double y);
+long double powx(long double x, int y);
 long double gamma(long double v);
 long double abs(long double *x, int size);
 long double abs(long double *x);
